Move test to mini-kevm
